requirements.txt
Django>=4.2
Pillow>=10.0.0
razorpay>=1.3.0
mysqlclient>=2.2.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2024.2
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
django-storages[google]>=1.14.2
python-dotenv>=1.0.0
firebase-admin>=6.5.0
